+++ src/de/ralleytn/simple/audio/HeadReader.java
@@ -1,0 +1,48 @@
+/*
+ * MIT License
+ * 
+ * Copyright (c) 2017 Ralph Niemitz
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package de.ralleytn.simple.audio;
+
+import java.io.IOException;
+import java.net.URL;
+import java.util.Map;
+
+/**
+ * Is implemented by all classes whith the purpose of reading the head informations from an audio file.
+ * @author Ralph Niemitz/RalleYTN(ralph.niemitz@gmx.de)
+ * @version 1.2.0
+ * @since 1.2.0
+ */
+public interface HeadReader {
+
+	/**
+	 * Reads the head informations of an audio file represented by an instance of {@linkplain URL}.<br>
+	 * <i>Warning:</i> This method does not check if an audio file is valid or not!
+	 * @param resource the instance of {@linkplain URL} representing the audio file
+	 * @return a {@linkplain Map} instance with all the readable head informations
+	 * @throws IOException if an I/O error occurs
+	 * @since 1.2.0
+	 */
+	public Map<String, ?> read(URL resource) throws IOException;
+}